Rename each repo folder to the full repo name, to simplify scripting
