+++ .gitmodules
@@ -1,18 +1,18 @@
 [submodule "root/termux-root-packages"]
-	path = root/termux-root-packages
+	path = termux-root-packages/termux-root-packages
 	url = https://github.com/termux/termux-root-packages
 [submodule "games/game-packages"]
-	path = games/game-packages
+	path = game-packages/game-packages
 	url = https://github.com/termux/game-packages
 [submodule "unstable/unstable-packages"]
-	path = unstable/unstable-packages
+	path = unstable-packages/unstable-packages
 	url = https://github.com/termux/unstable-packages
 [submodule "x11/x11-packages"]
-	path = x11/x11-packages
+	path = x11-packages/x11-packages
 	url = https://github.com/termux/x11-packages
 [submodule "science/science-packages"]
-	path = science/science-packages
+	path = science-packages/science-packages
 	url = https://github.com/termux/science-packages
 [submodule "main/termux-packages"]
-	path = main/termux-packages
+	path = termux-packages/termux-packages
 	url = https://github.com/termux/termux-packages
